# internal/api/server.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package api

import (
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
	"pull-request-api.com/internal/models"
	"pull-request-api.com/internal/service"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Создать PR и автоматически назначить до 2 ревьюверов из команды автора
	// (POST /pullRequest/create)
	PostPullRequestCreate(w http.ResponseWriter, r *http.Request)
	// Пометить PR как MERGED (идемпотентная операция)
	// (POST /pullRequest/merge)
	PostPullRequestMerge(w http.ResponseWriter, r *http.Request)
	// Переназначить конкретного ревьювера на другого из его команды
	// (POST /pullRequest/reassign)
	PostPullRequestReassign(w http.ResponseWriter, r *http.Request)
	// Создать команду с участниками (создаёт/обновляет пользователей)
	// (POST /team/add)
	PostTeamAdd(w http.ResponseWriter, r *http.Request)
	// Получить команду с участниками
	// (GET /team/get)
	GetTeamGet(w http.ResponseWriter, r *http.Request, params models.GetTeamGetParams)
	// Получить PR'ы, где пользователь назначен ревьювером
	// (GET /users/getReview)
	GetUsersGetReview(w http.ResponseWriter, r *http.Request, params models.GetUsersGetReviewParams)
	// Установить флаг активности пользователя
	// (POST /users/setIsActive)
	PostUsersSetIsActive(w http.ResponseWriter, r *http.Request)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.
type Server struct {
	ser *service.Service
}

func NewServer(ser *service.Service) *Server {
	return &Server{ser: ser}
}

// Создать PR и автоматически назначить до 2 ревьюверов из команды автора
// (POST /pullRequest/create)
func (s *Server) PostPullRequestCreate(w http.ResponseWriter, r *http.Request) {
	var body models.PostPullRequestCreateJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sendError(w, http.StatusBadRequest, models.NOTFOUND, "Invalid body")
		return
	}

	pr, err := s.ser.CreatePullRequest(r.Context(), body)
	if err != nil {
		handleServiceError(w, err)
		return
	}

	sendJSON(w, http.StatusOK, pr)
}

// Пометить PR как MERGED (идемпотентная операция)
// (POST /pullRequest/merge)
func (s *Server) PostPullRequestMerge(w http.ResponseWriter, r *http.Request) {
	var body models.PostPullRequestMergeJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sendError(w, http.StatusBadRequest, models.NOTFOUND, "Invalid body")
		return
	}

	pr, err := s.ser.MergePullRequest(r.Context(), body.PullRequestId)
	if err != nil {
		handleServiceError(w, err)
		return
	}

	sendJSON(w, http.StatusOK, pr)
}

// Переназначить конкретного ревьювера на другого из его команды
// (POST /pullRequest/reassign)
func (s *Server) PostPullRequestReassign(w http.ResponseWriter, r *http.Request) {
	var body models.PostPullRequestReassignJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sendError(w, http.StatusBadRequest, models.NOTFOUND, "Invalid body")
		return
	}

	pr, err := s.ser.ReassignReviewer(r.Context(), body)
	if err != nil {
		handleServiceError(w, err) //todo
		return
	}

	sendJSON(w, http.StatusOK, pr)
}

// Создать команду с участниками (создаёт/обновляет пользователей)
// (POST /team/add)
func (s *Server) PostTeamAdd(w http.ResponseWriter, r *http.Request) {
	var body models.Team
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sendError(w, http.StatusBadRequest, models.NOTFOUND, "Invalid body")
		return
	}

	err := s.ser.AddTeam(r.Context(), body)
	if err != nil {
		handleServiceError(w, err)
		return
	}

	sendJSON(w, http.StatusOK, body)
}

// Получить команду с участниками
// (GET /team/get)
func (s *Server) GetTeamGet(w http.ResponseWriter, r *http.Request, params models.GetTeamGetParams) {
	team, err := s.ser.GetTeam(r.Context(), params.TeamName)
	if err != nil {
		handleServiceError(w, err)
		return
	}
	sendJSON(w, http.StatusOK, team)
}

// Получить PR'ы, где пользователь назначен ревьювером
// (GET /users/getReview)
func (s *Server) GetUsersGetReview(w http.ResponseWriter, r *http.Request, params models.GetUsersGetReviewParams) {
	prs, err := s.ser.GetUsersReviews(r.Context(), params.UserId)
	if err != nil {
		handleServiceError(w, err)
		return
	}
	sendJSON(w, http.StatusOK, prs)
}

// Установить флаг активности пользователя
// (POST /users/setIsActive)
func (s *Server) PostUsersSetIsActive(w http.ResponseWriter, r *http.Request) {
	var body models.PostUsersSetIsActiveJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sendError(w, http.StatusBadRequest, models.NOTFOUND, "Invalid request body")
		return
	}

	user, err := s.ser.SetUserActive(r.Context(), body)
	if err != nil {
		handleServiceError(w, err)
		return
	}

	sendJSON(w, http.StatusOK, user)
}

func handleServiceError(w http.ResponseWriter, err error) {
	sendError(w, http.StatusNotFound, models.NOTFOUND, "Not found")
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// PostPullRequestCreate operation middleware
func (siw *ServerInterfaceWrapper) PostPullRequestCreate(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostPullRequestCreate(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PostPullRequestMerge operation middleware
func (siw *ServerInterfaceWrapper) PostPullRequestMerge(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostPullRequestMerge(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PostPullRequestReassign operation middleware
func (siw *ServerInterfaceWrapper) PostPullRequestReassign(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostPullRequestReassign(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PostTeamAdd operation middleware
func (siw *ServerInterfaceWrapper) PostTeamAdd(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostTeamAdd(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetTeamGet operation middleware
func (siw *ServerInterfaceWrapper) GetTeamGet(w http.ResponseWriter, r *http.Request) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params models.GetTeamGetParams

	// ------------- Required query parameter "team_name" -------------

	if paramValue := r.URL.Query().Get("team_name"); paramValue != "" {

	} else {
		siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "team_name"})
		return
	}

	err = runtime.BindQueryParameter("form", true, true, "team_name", r.URL.Query(), &params.TeamName)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "team_name", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetTeamGet(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetUsersGetReview operation middleware
func (siw *ServerInterfaceWrapper) GetUsersGetReview(w http.ResponseWriter, r *http.Request) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params models.GetUsersGetReviewParams

	// ------------- Required query parameter "user_id" -------------

	if paramValue := r.URL.Query().Get("user_id"); paramValue != "" {

	} else {
		siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "user_id"})
		return
	}

	err = runtime.BindQueryParameter("form", true, true, "user_id", r.URL.Query(), &params.UserId)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "user_id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetUsersGetReview(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PostUsersSetIsActive operation middleware
func (siw *ServerInterfaceWrapper) PostUsersSetIsActive(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostUsersSetIsActive(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/pullRequest/create", wrapper.PostPullRequestCreate)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/pullRequest/merge", wrapper.PostPullRequestMerge)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/pullRequest/reassign", wrapper.PostPullRequestReassign)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/team/add", wrapper.PostTeamAdd)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/team/get", wrapper.GetTeamGet)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/users/getReview", wrapper.GetUsersGetReview)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/users/setIsActive", wrapper.PostUsersSetIsActive)
	})

	return r
}